Rename CombinerCryptoClient and file
