+++ packages/phone-number-privacy/combiner/src/common/crypto-session.ts
@@ -4,7 +4,7 @@ import {
   SignMessageRequest,
 } from '@celo/phone-number-privacy-common'
 import { Request, Response } from 'express'
-import { CombinerCryptoClient } from './crypto-clients/crypto-client'
+import { CryptoClient } from './crypto-clients/common'
 import { Session } from './session'
 
 export class CryptoSession<
@@ -13,7 +13,7 @@ export class CryptoSession<
   public constructor(
     readonly request: Request<{}, {}, R>,
     readonly response: Response<OdisResponse<R>>,
-    readonly crypto: CombinerCryptoClient
+    readonly crypto: CryptoClient
   ) {
     super(request, response)
   }
